moved extension code to separate file (#28)
* moved extension code to separate file

* removed textfield delegate
